microservices/notification/requirements.txt
flask
flask-cors
kafka-python
requests
cryptography